refactor: contract verification contracts
