+++ typescript/sdk/src/deploy/verify/ContractVerifier.ts
@@ -1,5 +1,4 @@
 import fetch from 'cross-fetch';
-import { ethers } from 'ethers';
 import { Logger } from 'pino';
 
 import { rootLogger, sleep, strip0x } from '@hyperlane-xyz/utils';
@@ -8,6 +7,7 @@ import { ExplorerFamily } from '../../metadata/chainMetadataTypes.js';
 import { MultiProvider } from '../../providers/MultiProvider.js';
 import { ChainMap, ChainName } from '../../types.js';
 
+import { BaseContractVerifier } from './BaseContractVerifier.js';
 import {
   BuildArtifact,
   CompilerOptions,
@@ -19,12 +19,9 @@ import {
   SolidityStandardJsonInput,
 } from './types.js';
 
-export class ContractVerifier {
+export class ContractVerifier extends BaseContractVerifier {
   protected logger = rootLogger.child({ module: 'ContractVerifier' });
-
   protected contractSourceMap: { [contractName: string]: string } = {};
-
-  protected readonly standardInputJson: SolidityStandardJsonInput;
   protected readonly compilerOptions: CompilerOptions;
 
   constructor(
@@ -33,84 +30,117 @@ export class ContractVerifier {
     buildArtifact: BuildArtifact,
     licenseType: CompilerOptions['licenseType'],
   ) {
-    this.standardInputJson = buildArtifact.input;
-
+    super(multiProvider, buildArtifact);
     const compilerversion = `v${buildArtifact.solcLongVersion}`;
-
-    // double check compiler version matches expected format
     const versionRegex = /v(\d.\d.\d+)\+commit.\w+/;
     const matches = versionRegex.exec(compilerversion);
     if (!matches) {
       throw new Error(`Invalid compiler version ${compilerversion}`);
     }
-
-    // set compiler options
-    // only license type is configurable, empty if not provided
     this.compilerOptions = {
       codeformat: 'solidity-standard-json-input',
       compilerversion,
       licenseType,
     };
+  }
 
-    // process input to create mapping of contract names to source names
-    // this is required to construct the fully qualified contract name
-    const contractRegex = /contract\s+([A-Z][a-zA-Z0-9]*)/g;
-    Object.entries(buildArtifact.input.sources).forEach(
-      ([sourceName, { content }]) => {
-        const matches = content.matchAll(contractRegex);
-        for (const match of matches) {
-          const contractName = match[1];
-          if (contractName) {
-            this.contractSourceMap[contractName] = sourceName;
-          }
-        }
+  protected async verify(
+    chain: ChainName,
+    input: ContractVerificationInput,
+    verificationLogger: Logger,
+  ): Promise<void> {
+    const contractType: string = input.isProxy ? 'proxy' : 'implementation';
+
+    verificationLogger.debug(`📝 Verifying ${contractType}...`);
+
+    const data = input.isProxy
+      ? this.getProxyData(input)
+      : this.getImplementationData(chain, input, verificationLogger);
+
+    try {
+      const guid: string = await this.submitForm(
+        chain,
+        input.isProxy
+          ? ExplorerApiActions.VERIFY_PROXY
+          : ExplorerApiActions.VERIFY_IMPLEMENTATION,
+        verificationLogger,
+        data,
+      );
+
+      verificationLogger.trace(
+        { guid },
+        `Retrieved guid from verified ${contractType}.`,
+      );
+
+      await this.checkStatus(
+        chain,
+        input,
+        verificationLogger,
+        guid,
+        contractType,
+      );
+
+      const addressUrl = await this.multiProvider.tryGetExplorerAddressUrl(
+        chain,
+        input.address,
+      );
+
+      verificationLogger.debug(
+        {
+          addressUrl: addressUrl
+            ? `${addressUrl}#code`
+            : `Could not retrieve ${contractType} explorer URL.`,
+        },
+        `✅ Successfully verified ${contractType}.`,
+      );
+    } catch (error) {
+      verificationLogger.debug(
+        { error },
+        `Verification of ${contractType} failed`,
+      );
+      throw error;
+    }
+  }
+
+  private async checkStatus(
+    chain: ChainName,
+    input: ContractVerificationInput,
+    verificationLogger: Logger,
+    guid: string,
+    contractType: string,
+  ): Promise<void> {
+    verificationLogger.trace({ guid }, `Checking ${contractType} status...`);
+    await this.submitForm(
+      chain,
+      input.isProxy
+        ? ExplorerApiActions.CHECK_PROXY_STATUS
+        : ExplorerApiActions.CHECK_IMPLEMENTATION_STATUS,
+      verificationLogger,
+      {
+        guid: guid,
       },
     );
   }
 
-  public async verifyContract(
-    chain: ChainName,
-    input: ContractVerificationInput,
-    logger = this.logger,
-  ): Promise<void> {
-    const verificationLogger = logger.child({
-      chain,
-      name: input.name,
+  private getProxyData(input: ContractVerificationInput) {
+    return {
       address: input.address,
-    });
+      expectedimplementation: input.expectedimplementation,
+    };
+  }
 
-    const metadata = this.multiProvider.tryGetChainMetadata(chain);
-    const rpcUrl = metadata?.rpcUrls[0].http ?? '';
-    if (rpcUrl.includes('localhost') || rpcUrl.includes('127.0.0.1')) {
-      verificationLogger.debug('Skipping verification for local endpoints');
-      return;
-    }
-
-    const explorerApi = this.multiProvider.tryGetExplorerApi(chain);
-    if (!explorerApi) {
-      verificationLogger.debug('No explorer API set, skipping');
-      return;
-    }
-
-    if (!explorerApi.family) {
-      verificationLogger.debug(`No explorer family set, skipping`);
-      return;
-    }
-
-    if (explorerApi.family === ExplorerFamily.Other) {
-      verificationLogger.debug(`Unsupported explorer family, skipping`);
-      return;
-    }
-
-    if (input.address === ethers.constants.AddressZero) return;
-    if (Array.isArray(input.constructorArguments)) {
-      verificationLogger.debug(
-        'Constructor arguments in legacy format, skipping',
-      );
-      return;
-    }
-
-    await this.verify(chain, input, verificationLogger);
+  protected prepareImplementationData(
+    sourceName: string,
+    input: ContractVerificationInput,
+    filteredStandardInputJson: SolidityStandardJsonInput,
+  ) {
+    return {
+      sourceCode: JSON.stringify(filteredStandardInputJson),
+      contractname: `${sourceName}:${input.name}`,
+      contractaddress: input.address,
+      constructorArguements: strip0x(input.constructorArguments ?? ''),
+      ...this.compilerOptions,
+    };
   }
 
   private async submitForm(
@@ -255,204 +285,5 @@ export class ContractVerifier {
 
     await sleep(timeout);
     return responseJson.result;
-  }
-
-  private async verify(
-    chain: ChainName,
-    input: ContractVerificationInput,
-    verificationLogger: Logger,
-  ): Promise<void> {
-    const contractType: string = input.isProxy ? 'proxy' : 'implementation';
-
-    verificationLogger.debug(`📝 Verifying ${contractType}...`);
-
-    const data = input.isProxy
-      ? this.getProxyData(input)
-      : this.getImplementationData(chain, input, verificationLogger);
-
-    try {
-      const guid: string = await this.submitForm(
-        chain,
-        input.isProxy
-          ? ExplorerApiActions.VERIFY_PROXY
-          : ExplorerApiActions.VERIFY_IMPLEMENTATION,
-        verificationLogger,
-        data,
-      );
-
-      verificationLogger.trace(
-        { guid },
-        `Retrieved guid from verified ${contractType}.`,
-      );
-
-      await this.checkStatus(
-        chain,
-        input,
-        verificationLogger,
-        guid,
-        contractType,
-      );
-
-      const addressUrl = await this.multiProvider.tryGetExplorerAddressUrl(
-        chain,
-        input.address,
-      );
-
-      verificationLogger.debug(
-        {
-          addressUrl: addressUrl
-            ? `${addressUrl}#code`
-            : `Could not retrieve ${contractType} explorer URL.`,
-        },
-        `✅ Successfully verified ${contractType}.`,
-      );
-    } catch (error) {
-      verificationLogger.debug(
-        { error },
-        `Verification of ${contractType} failed`,
-      );
-      throw error;
-    }
-  }
-
-  private async checkStatus(
-    chain: ChainName,
-    input: ContractVerificationInput,
-    verificationLogger: Logger,
-    guid: string,
-    contractType: string,
-  ): Promise<void> {
-    verificationLogger.trace({ guid }, `Checking ${contractType} status...`);
-    await this.submitForm(
-      chain,
-      input.isProxy
-        ? ExplorerApiActions.CHECK_PROXY_STATUS
-        : ExplorerApiActions.CHECK_IMPLEMENTATION_STATUS,
-      verificationLogger,
-      {
-        guid: guid,
-      },
-    );
-  }
-
-  private getProxyData(input: ContractVerificationInput) {
-    return {
-      address: input.address,
-      expectedimplementation: input.expectedimplementation,
-    };
-  }
-
-  private getImplementationData(
-    chain: ChainName,
-    input: ContractVerificationInput,
-    verificationLogger: Logger,
-  ) {
-    const sourceName = this.contractSourceMap[input.name];
-    if (!sourceName) {
-      const errorMessage = `Contract '${input.name}' not found in provided build artifact`;
-      verificationLogger.error(errorMessage);
-      throw new Error(`[${chain}] ${errorMessage}`);
-    }
-
-    const filteredStandardInputJson =
-      this.filterStandardInputJsonByContractName(
-        input.name,
-        this.standardInputJson,
-        verificationLogger,
-      );
-
-    return {
-      sourceCode: JSON.stringify(filteredStandardInputJson),
-      contractname: `${sourceName}:${input.name}`,
-      contractaddress: input.address,
-      /* TYPO IS ENFORCED BY API */
-      constructorArguements: strip0x(input.constructorArguments ?? ''),
-      ...this.compilerOptions,
-    };
-  }
-
-  /**
-   * Filters the solidity standard input for a specific contract name.
-   *
-   * This is a BFS impl to traverse the source input dependency graph.
-   * 1. Named contract file is set as root node.
-   * 2. The next level is formed by the direct imports of the contract file.
-   * 3. Each subsequent level's dependencies form the next level, etc.
-   * 4. The queue tracks the next files to process, and ensures the dependency graph explorered level by level.
-   */
-  private filterStandardInputJsonByContractName(
-    contractName: string,
-    input: SolidityStandardJsonInput,
-    verificationLogger: Logger,
-  ): SolidityStandardJsonInput {
-    verificationLogger.trace(
-      { contractName },
-      'Filtering unused contracts from solidity standard input JSON....',
-    );
-    const filteredSources: SolidityStandardJsonInput['sources'] = {};
-    const sourceFiles: string[] = Object.keys(input.sources);
-    const contractFile: string = this.getContractFile(
-      contractName,
-      sourceFiles,
-    );
-    const queue: string[] = [contractFile];
-    const processed = new Set<string>();
-
-    while (queue.length > 0) {
-      const file = queue.shift()!;
-      if (processed.has(file)) continue;
-      processed.add(file);
-
-      filteredSources[file] = input.sources[file];
-
-      const content = input.sources[file].content;
-      const importStatements = this.getAllImportStatements(content);
-
-      importStatements.forEach((importStatement) => {
-        const importPath = importStatement.match(/["']([^"']+)["']/)?.[1];
-        if (importPath) {
-          const resolvedPath = this.resolveImportPath(file, importPath);
-          if (sourceFiles.includes(resolvedPath)) queue.push(resolvedPath);
-        }
-      });
-    }
-
-    return {
-      ...input,
-      sources: filteredSources,
-    };
-  }
-
-  private getContractFile(contractName: string, sourceFiles: string[]): string {
-    const contractFile = sourceFiles.find((file) =>
-      file.endsWith(`/${contractName}.sol`),
-    );
-    if (!contractFile) {
-      throw new Error(`Contract ${contractName} not found in sources.`);
-    }
-    return contractFile;
-  }
-
-  private getAllImportStatements(content: string) {
-    const importRegex =
-      /import\s+(?:(?:(?:"[^"]+"|'[^']+')\s*;)|(?:{[^}]+}\s+from\s+(?:"[^"]+"|'[^']+')\s*;)|(?:\s*(?:"[^"]+"|'[^']+')\s*;))/g;
-    return content.match(importRegex) || [];
-  }
-
-  private resolveImportPath(currentFile: string, importPath: string): string {
-    /* Use as-is for external dependencies and absolute imports */
-    if (importPath.startsWith('@') || importPath.startsWith('http')) {
-      return importPath;
-    }
-    const currentDir = currentFile.split('/').slice(0, -1).join('/');
-    const resolvedPath = importPath.split('/').reduce((acc, part) => {
-      if (part === '..') {
-        acc.pop();
-      } else if (part !== '.') {
-        acc.push(part);
-      }
-      return acc;
-    }, currentDir.split('/'));
-    return resolvedPath.join('/');
   }
 }
